+++ backend/src/models/Organisation.js
@@ -10,10 +10,9 @@ class Organisation extends Model {
   static get jsonSchema() {
     return {
       type: 'object',
-      required: ['org_id'],
+      required: ['org_name'],
 
       properties: {
-        org_id: { type: 'integer' },
         org_name: { type: 'string', minLength: 1, maxLength: 255 },
         website: { type: 'string', minLength: 1, maxLength: 255 },
         email_address: { type: 'string', minLength: 1, maxLength: 255 },
